Move bytes and hex from revive to the types crate
