+++ substrate/frame/revive/types/src/common/mod.rs
@@ -15,6 +15,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod byte;
 mod serde_helpers;
 
+pub use byte::*;
 pub use serde_helpers::*;
