+++ substrate/frame/revive/types/src/lib.rs
@@ -17,5 +17,7 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+extern crate alloc;
+
 pub mod common;
 pub mod runtime_api;
